Added default chattrbot namespace. Now we just need to use the global styles from the tailwind config
