--- src/chatbots/index.ts
@@ -1,2 +1,0 @@
-export { default as ChatBot } from './chatbot'
-export { default as ChattrBot } from './chattrbot'
