+++ src/components/default/chattrbot.tsx
@@ -1,47 +1,33 @@
 import React from 'react'
 
-import { useChattr } from '../hooks'
+import { useChattr } from '../../hooks'
 
-import {
-  ChattrContainer,
-  ChattrFeed,
-  ChattrForm,
-  ChattrHeader,
-  ChattrIcon,
-} from '../components/minimalist'
+import ChattrContainer from './chattr-container'
+import ChattrFeed from './chattr-feed'
+import ChattrForm from './chattr-form'
+import ChattrHeader from './chattr-header'
+import ChattrIcon from './chattr-icon'
 
-type ChattrMessage = {
+type ChattrMessages = {
   role: 'user' | 'assistant'
   key?: string
   content?: string
-  ui: string
-  data?: {
-    temperature: string
-    celcius: string
-    location: string
-    description: string
-    humidity: string
-    wind: string
-    clouds: string
-    state: string
-    url: string
-  }
 }
 
 /**
  * Chattrbot - A pre made chatbot solution with light/dark mode tailwind css classes.
  *
- * @returns The entire Chattrbot component.
+ * @returns The entire Default Chattrbot component.
  */
 
 export default function Chattrbot() {
   const { isOpen, toggle } = useChattr()
-  const chattrBotName = 'Chattrbot'
+  const chattrBotName = 'Bot'
+  const userName = 'User'
   const [loading, setLoading] = React.useState(false)
   const [message, setMessage] = React.useState('')
-  const [messages, setMessages] = React.useState<ChattrMessage[]>([
+  const [messages, setMessages] = React.useState<ChattrMessages[]>([
     {
-      ui: 'default',
       content: `Hey! Thanks for visiting. I'm ${chattrBotName}, you can ask me anything!`,
       role: 'assistant',
     },
@@ -56,11 +42,11 @@ export default function Chattrbot() {
     setMessages(prev => [
       ...prev,
       {
-        ui: 'default',
         content: message,
         role: 'user',
       },
     ])
+
     const response = await fetch('/api/chat-gpt', {
       method: 'POST',
       headers: {
@@ -81,44 +67,33 @@ export default function Chattrbot() {
         setMessages(prev => [
           ...prev,
           {
-            ui: 'default',
             content: result.content.text,
             role: 'assistant',
           },
         ])
-      } else if (result.content.function_response) {
+      } else {
         setLoading(false)
         setMessages(prev => [
           ...prev,
           {
-            ui: result.ui,
-            data: result.content.function_response,
+            content: result.error,
             role: 'assistant',
           },
         ])
       }
-    } else {
-      setLoading(false)
-      setMessages(prev => [
-        ...prev,
-        {
-          ui: 'default',
-          content: result.error,
-          role: 'assistant',
-        },
-      ])
     }
   }
 
   return isOpen ? (
     <ChattrContainer>
       <ChattrHeader
-        chattrBotName={chattrBotName}
-        role='Assistant'
         toggle={toggle}
+        chattrBotName={chattrBotName}
       />
       <ChattrFeed
         messages={messages}
+        userName={userName}
+        chattrBotName={chattrBotName}
         loading={loading}
       />
       <ChattrForm
