+++ src/components/default/index.ts
@@ -1,7 +1,10 @@
-export { default as ChatContainer } from './chat-container'
-export { default as ChatFeed } from './chat-feed'
-export { default as ChatHeader } from './chat-header'
-export { default as ChatIcon } from './chat-icon'
-export { default as ChatInput } from './chat-input'
-export { default as ChatLoader } from './chat-loader'
-export { default as ChatMessage } from './chat-message'
+export { default as ChattrAssistantMessage } from './chattr-assistant-message'
+export { default as ChattrUserMessage } from './chattr-user-message'
+export { default as ChattrContainer } from './chattr-container'
+export { default as ChattrFeed } from './chattr-feed'
+export { default as ChattrForm } from './chattr-form'
+export { default as ChattrHeader } from './chattr-header'
+export { default as ChattrIcon } from './chattr-icon'
+export { default as ChattrLoader } from './chattr-loader'
+export { default as ChattrMessages } from './chattr-messages'
+export { default as Chattrbot } from './chattrbot'
