+++ src/components/minimalist/index.ts
@@ -10,3 +10,4 @@ export { default as ChattrHeader } from './chattr-header'
 export { default as ChattrIcon } from './chattr-icon'
 export { default as ChattrLoader } from './chattr-loader'
 export { default as ChattrMessages } from './chattr-messages'
+export { default as Chattrbot } from './chattrbot'
